pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eidolon-system"
version = "0.1.0"
description = "Large robot fleet management system"
authors = [{name = "Eidolon Team", email = "team@eidolon.cloud"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "grpcio>=1.59.3",
    "grpcio-tools>=1.59.3",
    "protobuf>=4.25.1",
    "pydantic>=2.5.0",
    "aiortc>=1.6.0",
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
    "redis>=5.0.1",
    "boto3>=1.34.0",
    "kafka-python>=2.0.2",
    "cryptography>=41.0.8",
    "pyjwt>=2.8.0",
    "torch>=2.1.1",
    "opencv-python>=4.8.1.78",
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
